fix: explicit float16 type def

+++ models/encoders.py
@@ -22,8 +22,8 @@ class ClassConditionNorm(nn.Module):
         self.eps = eps
         self.norm_style = norm_style 
 
-        self.register_buffer('stored_mean', torch.zeros(output_size).to(torch.float16))
-        self.register_buffer('stored_var', torch.ones(output_size).to(torch.float16)) 
+        self.register_buffer('stored_mean', torch.zeros(output_size))
+        self.register_buffer('stored_var', torch.ones(output_size)) 
 
     def forward(self, x, y):
         # Calculate class-conditional gains and biases
